+++ CHANGELOG.md
@@ -6,6 +6,7 @@ This is a major refactor!
 * awesn1 instead of indispensable-asn1
 * publish javadoc redirect to save space
 * make all provider functions suspend
+* Replace `JwsHeader.Part` with `JwsHeader.unprotectedMembers`-based header splitting
 * Dependency Updates:
     * Bouncy Castle 1.85
     * multibase 1.3.0